+++ R/hypotheses.R
@@ -166,7 +166,11 @@ hypotheses <- function(
     if (inherits(model, c("predictions", "comparisons", "slopes", "hypotheses"))) {
         if (!is.null(vcov)) {
             msg <- "The `vcov` argument is not available when `model` is a `predictions`, `comparisons`, `slopes`, or `hypotheses` object. Please specify the type of standard errors in the initial `marginaleffects` call."
-            stop(msg, call. = FALSE)
+            stop_sprintf(msg)
+        }
+        if (!is.null(attr(model, "posterior_draws"))) {
+            msg <- "The `hypotheses()` function cannot be used to post-process `marginaleffects` objects that include draws from a bootstrap, simulation, or bayesian posterior distribution. Users should call the `get_draws()` function and process draws manually."
+            stop_sprintf(msg)
         }
     }